Fix a GC issue when creating datobjs
Calling SET_TYPE_DATOBJ only writes a pointer into the object, but does not
call CHANGED_BAG; so for types that are created but not reference elsewhere,
there is a minor chance of a GC bug here.

We fix this in two ways: First, by ensuring the type is created *before* the
object, and thus never younger. Second, by calling SET_TYPE_OBJ instead, which
also calls CHANGED_BAG, and as a bonus, may improve things on HPC-GAP.